+++ tdms/src/hdf5_io/hdf5_base.cpp
@@ -1,20 +1,18 @@
-#include "hdf5_io.h"
-#include "cell_coordinate.h"
+/**
+ * @file hdf5_io.cpp
+ * @authors Sam Cunliffe, William Graham
+ * @brief Common HDF5 I/O methods abstracted to the base class.
+ */
+#include "hdf5_io/hdf5_base.h"
 
-#include <H5public.h>
 #include <iostream>
 
 #include <H5Cpp.h>
+#include <H5public.h>
 #include <spdlog/spdlog.h>
 
-/**
- * @brief Convert from a vector of HDF5's hsize_t back to our struct of ints.
- * @note Local scope utility function as only this code needs to interact with
- * the HDF5 H5Cpp library.
- *
- * @param dimensions a 1, 2, or 3 element vector of dimensions.
- * @return ijk The dimensions in a struct.
- */
+using namespace std;
+
 ijk to_ijk(const std::vector<hsize_t> dimensions) {
   unsigned int rank = dimensions.size();
   ijk out;
@@ -25,30 +23,8 @@ ijk to_ijk(const std::vector<hsize_t> dimensions) {
   return out;
 }
 
-/******************************************************************************
- * HDF5Writer
- */
-void HDF5Writer::write(const std::string &dataset_name, double *data, int size,
-                       hsize_t *dimensions) {
-  spdlog::debug("Writing {} to file: {}", dataset_name, filename_);
-
-  // declare a dataspace
-  H5::DataSpace dataspace(size, dimensions);
-  H5::DataType datatype(H5::PredType::NATIVE_DOUBLE);
-
-  // write the data to the dataset object in the file
-  H5::DataSet dataset = file_->createDataSet(dataset_name, datatype, dataspace);
-  dataset.write(data, H5::PredType::NATIVE_DOUBLE);
-  spdlog::trace("Write successful.");
-}
-
-/******************************************************************************
- * HDF5Base
- *
- * Common HDF5 I/O methods abstracted to the base class.
- */
-std::vector<std::string> HDF5Base::get_datanames() const {
-  std::vector<std::string> names;
+vector<string> HDF5Base::get_datanames() const {
+  vector<string> names;
 
   // iterate over all objects in the file
   for (unsigned int i = 0; i < file_->getNumObjs(); i++) {
@@ -64,14 +40,14 @@ std::vector<std::string> HDF5Base::get_datanames() const {
 }
 
 void HDF5Base::ls() const {
-  std::vector<std::string> names = this->get_datanames();
-  for (auto name : names) std::cout << name << std::endl;
+  vector<string> names = this->get_datanames();
+  for (auto name : names) cout << name << endl;
   return;
 }
 
 // IJKDimensions HDF5Base::shape_of(const std::string &dataname) const {
 // return to_ijk(dimensions);
-std::vector<hsize_t> HDF5Base::shape_of(const std::string &dataname) const {
+vector<hsize_t> HDF5Base::shape_of(const string &dataname) const {
   SPDLOG_DEBUG("shape_of");
 
   // get the dataset and dataspace (contains dimensionality info)
@@ -80,7 +56,7 @@ std::vector<hsize_t> HDF5Base::shape_of(const std::string &dataname) const {
 
   // need the rank in order to declare the vector size
   int rank = dataspace.getSimpleExtentNdims();
-  std::vector<hsize_t> dimensions(rank);
+  vector<hsize_t> dimensions(rank);
   dataspace.getSimpleExtentDims(dimensions.data(), nullptr);
   return dimensions;
 }
