+++ tdms/tests/unit/hdf5_io_tests/structure_array.mat
